llpc_project/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-countries==7.6.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
idna==3.10
mysqlclient==2.2.7
pillow==11.2.1
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0